Unify duplicated code introduced into AudioWidget and VideoWidget
